+++ recipes/oozie.rb
@@ -45,7 +45,7 @@ java_share_dir = '/usr/share/java'
 case node['platform_family']
 when 'debian'
   pkgs = %w(
-    libmysql-java
+    mysql-connector-java
     libpostgresql-jdbc-java
   )
   jars = %w(
@@ -82,7 +82,7 @@ package 'unzip'
 
 extjs = 'ext-2.2.zip'
 remote_file "#{oozie_data_dir}/#{extjs}" do
-  source "http://extjs.com/deploy/#{extjs}"
+  source "http://dev.sencha.com/deploy/#{extjs}"
   mode '0644'
   action :create_if_missing
 end