+++ spec/spark_historyserver_spec.rb
@@ -39,7 +39,7 @@ describe 'hadoop::spark_historyserver' do
       ChefSpec::SoloRunner.new(platform: 'centos', version: 6.5) do |node|
         node.automatic['domain'] = 'example.com'
         node.override['hadoop']['distribution'] = 'cdh'
-        node.override['hadoop']['distribution_version'] = 5
+        node.override['hadoop']['distribution_version'] = '5.3.2'
         stub_command('test -L /var/log/spark').and_return(false)
         stub_command('update-alternatives --display spark-conf | grep best | awk \'{print $5}\' | grep /etc/spark/conf.chef').and_return(false)
       end.converge(described_recipe)